+++ gitman/__init__.py
@@ -3,13 +3,13 @@
 import sys
 
 __project__ = 'GitMan'
-__version__ = '1.0.2'
+__version__ = '1.1'
 
 CLI = 'gitman'
 PLUGIN = 'deps'
-NAME = "Git Dependency Manager"
+NAME = 'Git Dependency Manager'
 VERSION = __project__ + ' v' + __version__
-DESCRIPTION = "A language-agnostic dependency manager using Git."
+DESCRIPTION = 'A language-agnostic dependency manager using Git.'
 
 PYTHON_VERSION = 3, 5