+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision History
 
+## 1.1 (2017/01/06)
+
+- Added coloring to the command-line output.
+- Added preliminary Windows support.
+- Fixed issue where `<dirty>` could be saved as a locked revision.
+
 ## 1.0.2 (2016/07/28)
 
 - Moved documentation to http://gitman.readthedocs.io/.